requirements.txt
absl-py==2.3.1
affine @ file:///home/conda/feedstock_root/build_artifacts/affine_1733762038348/work
asf_search==11.0.0
astunparse==1.6.3
attrs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_attrs_1764875182/work
cattrs==25.3.0
certifi @ file:///home/conda/feedstock_root/build_artifacts/certifi_1767500808759/work/certifi
charset-normalizer==3.4.4
click @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_click_1764518652/work
click-plugins @ file:///home/conda/feedstock_root/build_artifacts/click-plugins_1750848229740/work
cligj @ file:///home/conda/feedstock_root/build_artifacts/cligj_1733749956636/work
contourpy==1.3.2
cycler==0.12.1
dateparser==1.2.2
dill==0.4.0
exceptiongroup==1.3.1
flatbuffers==25.9.23
fonttools==4.61.1
gast==0.4.0
GDAL==3.5.3
GeoAlchemy2==0.13.3
geopandas==1.0.1
google-auth==2.47.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.3.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
idna==3.11
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_importlib-metadata_1747934053/work
jax @ file:///home/conda/feedstock_root/build_artifacts/jax_1733731203765/work
jaxlib @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_2a01rjwprl/croot/jaxlib_1732663329121/work/dist/jaxlib-0.4.35-cp310-cp310-macosx_10_14_x86_64.whl#sha256=4b5f1f56e0cd1dfbd33b7de31d56edb41671d612000d9a27ceb9f41c69c92918
jh2==5.0.10
joblib==1.5.3
keras==2.12.0
kiwisolver==1.4.9
libclang==18.1.1
lxml==6.0.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
ml_dtypes @ file:///Users/runner/miniforge3/conda-bld/bld/rattler-build_ml_dtypes_1764068383/work
multiprocess==0.70.18
niquests==3.16.1
numpy==1.23.5
oauthlib==3.3.1
openmeteo_requests==1.7.4
openmeteo_sdk==1.24.0
opt_einsum @ file:///home/conda/feedstock_root/build_artifacts/opt_einsum_1733687912731/work
packaging==25.0
pandas==2.3.3
pathos==0.3.4
pillow==12.1.0
platformdirs==4.5.1
pox==0.3.6
ppft==1.7.7
progressbar2==4.5.0
protobuf==4.25.8
psycopg2 @ file:///Users/runner/miniforge3/conda-bld/psycopg2-split_1672159344532/work
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyogrio==0.12.1
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1766475444/work
pyproj==3.7.1
pyroSAR==0.33.0
python-dateutil==2.9.0.post0
python-utils==3.9.1
pytz==2025.2
PyYAML==6.0.3
qh3==1.5.6
rasterio @ file:///Users/runner/miniforge3/conda-bld/rasterio_1667824043189/work
regex==2025.11.3
requests==2.32.5
requests-cache==1.2.1
requests-oauthlib==2.0.0
retry-requests==2.0.0
rsa==4.9.1
scikit-learn==1.7.2
scipy @ file:///Users/runner/miniforge3/conda-bld/scipy-split_1739790744051/work/dist/scipy-1.15.2-cp310-cp310-macosx_10_13_x86_64.whl#sha256=f6d40aea21a34e9d9a143f479557cad14bca2f4dbd51ceb15ddc40e976676acb
Shapely @ file:///Users/runner/miniforge3/conda-bld/shapely_1668517998851/work
six==1.17.0
snuggs @ file:///home/conda/feedstock_root/build_artifacts/snuggs_1733818638588/work
spatialist==0.16.1
SQLAlchemy==1.4.54
SQLAlchemy-Utils==0.41.2
tblib==3.2.2
tenacity==9.1.2
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.12.0
tensorflow-metal==1.0.0
termcolor==3.3.0
threadpoolctl==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
url-normalize==2.2.1
urllib3==2.6.3
urllib3-future==2.15.901
wassima==2.0.4
Werkzeug==3.1.5
wrapt==1.14.2
zipp @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_zipp_1764460141/work
